move auth related components to src/auth
